@@ -16,44 +16,47 @@
     "checks": "npm run lint && npm run test"
   },
   "devDependencies": {
-    "autoprefixer": "^6.5.0",
+    "autoprefixer": "^7.1.5",
     "babel-core": "^6.17.0",
+    "babel-eslint": "^8.0.1",
     "babel-preset-es2015": "^6.16.0",
     "babel-preset-react": "^6.16.0",
     "babel-register": "^6.16.3",
     "babelify": "^7.3.0",
-    "browserify": "^13.1.0",
-    "cross-env": "^3.1.3",
-    "envify": "^4.0.0",
-    "enzyme": "^2.4.1",
+    "browserify": "^14.4.0",
+    "cross-env": "^5.0.5",
+    "envify": "^4.1.0",
+    "enzyme": "^3.1.0",
+    "enzyme-adapter-react-15": "^1.0.1",
     "eslint": "^3.7.1",
     "eslint-config-bitcrowd-react": "^12.0.0",
     "eslint-plugin-import": "^1.16.0",
     "eslint-plugin-jsx-a11y": "^2.2.3",
     "eslint-plugin-react": "^6.4.0",
     "gulp": "^3.9.1",
-    "gulp-postcss": "^6.2.0",
-    "gulp-sass": "^2.3.2",
-    "gulp-sourcemaps": "^2.0.0",
-    "gulp-uglify": "^2.0.0",
-    "gulp-zip": "^3.2.0",
+    "gulp-postcss": "^7.0.0",
+    "gulp-sass": "^3.1.0",
+    "gulp-sourcemaps": "^2.6.1",
+    "gulp-uglify": "^3.0.0",
+    "gulp-zip": "^4.0.0",
     "jasmine": "^2.5.2",
     "jsdom": "^11.3.0",
-    "nodemon": "^1.11.0",
+    "nodemon": "^1.12.1",
     "nyc": "^10.0.0",
-    "react-addons-test-utils": "^15.3.2",
-    "rimraf": "^2.5.4",
-    "through2": "^2.0.1",
+    "react-test-renderer": "^15.6.1",
+    "rimraf": "^2.6.2",
+    "through2": "^2.0.3",
     "vinyl-buffer": "^1.0.0",
     "vinyl-source-stream": "^1.1.0"
   },
   "dependencies": {
     "bootstrap": "=4.0.0-alpha.4",
-    "jquery": "^3.1.1",
-    "octicons": "^5.0.1",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "jquery": "^3.2.1",
+    "octicons": "^7.0.0",
+    "prop-types": "^15.6.0",
+    "react": "^15.6.1",
+    "react-dom": "^15.6.1",
     "react-router": "^3.0.0",
-    "speakingurl": "^10.0.0"
+    "speakingurl": "^14.0.1"
   }
 }
